Interface for setting external offscreen renderer functionality:
Makes it possible to provide an external implementation for doing
offscreen rendering (such as accelerated rendering into a hidden
window). This is useful to avoid having to do slow software
rendering when pBuffers cannot be used (not available on the
system, buggy implementation, mismatch of onscreen context and
offscreen context).