Killed 087:
This was not a "real" bug: the program didn't hang, it just tried to
use insane amounts of memory because the person designing that
specific curve had chosen a far-out high resolution in the v
domain.

Changing v from [0; 2388.08] to [0; 2.38808] lets me open the file
just fine.

NURBS -- not for the faint of heart, or the uninitiated.